[OU-IMP] website_event: Add script to insert default event questions for old events
In version 16, the default questions in the registration form, such as
‘Name’, ‘Email’ and ‘Phone’, were directly visible in the registration
form view for users. However, in Odoo v17, although those questions are
automatically created in new events as records in the event.question
template, they are not inserted in existing events, causing a lack of
consistency in the registration form.
This script solves that problem by adding those questions by default
to all existing events, ensuring that they have the same behaviour as
new events. For the translations, instead of defining them manually,
the script takes advantage of existing translations in the ir.model.fields
table for the name, email and phone fields of the res.partner model.

TT51546